Bugfix/user username realname fix #156

+++ main.cpp
@@ -2,10 +2,10 @@
 
 int main(int argc, char* argv[])
 {
-	if (argc != 3 || atoi(argv[1]) == -1 || strlen(argv[2]) == 0)
+	if (argc != 3 || atol(argv[1]) == -1 || strlen(argv[2]) == 0)
 		return (1);
 	
-	Server server(atoi(argv[1]), argv[2]);
+	Server server(atol(argv[1]), argv[2]);
 	server.run();
 
 	return (0);
